test_mgmt/requirements.txt
setuptools

APScheduler==3.11.0
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.1
customtkinter==5.2.2
darkdetect==0.8.0
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2.6
django-ace==1.43.3
django-allauth==65.11.2
django-cors-headers==4.9.0
django-crontab==0.7.1
django-dbbackup==5.0.0
django-env-overrides==0.1.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-import-export==4.3.9
django-jazzmin==3.0.1
django-json-field==0.5.7
django-jsonfield==1.4.1
django-mass-edit==3.5.0
django_yaml_field==0.1.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupPy==1.18
MarkupSafe==3.0.2
numpy==2.2.6
oauthlib==3.3.1
odfpy==1.4.1
openapi-codec==1.3.2
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
Python_IO==0.3
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
requests==2.32.5
requests-oauthlib==2.0.0
scipy==1.16.2

simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
xlrd==2.0.2
xlwt==1.3.0
